requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
celery==5.3.4
redis==5.0.1
mysql-connector-python==8.2.0
pydantic==2.4.2
python-dotenv==1.0.0
SQLAlchemy==2.0.23
alembic==1.12.1
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
requests==2.31.0
aiohttp==3.8.6
numpy==1.26.1
pandas==2.1.2
matplotlib==3.8.1